Rename loadingStatusService to loadingStatus.service #634
