+++ visualization/app/codeCharta/state/state.module.ts
@@ -6,7 +6,7 @@ import { NodeSearchService } from "./nodeSearch.service"
 import { FileStateService } from "./fileState.service"
 import { SettingsService } from "./settingsService/settings.service"
 import { MetricService } from "./metric.service"
-import { LoadingStatusService } from "./loadingStatusService"
+import { LoadingStatusService } from "./loadingStatus.service"
 import "../ui/loadingGif/loadingGif.module"
 import "../codeCharta.module"
 import _ from "lodash"
